add AdamOptimizer, fix datatype (#1683)
* add AdamOptimizer, fix datatype

* add AdamOptimizer test, fix datatype. TODO: implement sparseAdam

* add sparseAdam, fix some datatypes and other funtions

* add missing files

* fix other test file

* add SparseAdam unittest

* add tensor function

* tiny fixes

* tiny fixes

+++ elasticdl/pkg/kernel/capi/kernel_api.cc
@@ -17,12 +17,12 @@ void Adam(float* grad,
           float* param,
           float* m,
           float* v,
-          double lr,
+          float lr,
           long long size,
           long long step,
-          double beta1,
-          double beta2,
-          double epsilon,
+          float beta1,
+          float beta2,
+          float epsilon,
           float* max_square) {
   Eigen::Map<Eigen::Array<float, 1, Eigen::Dynamic>> eg{
       grad, static_cast<Eigen::Index>(size)};
